requirements.txt
numpy
pillow
openai
torch